+++ files/XlsxFileType.ts
@@ -1,4 +1,4 @@
-﻿/** Information about an XLSX file
+﻿/** Metadata about an XLSX file
  * @author TeamworkGuy2
  * @since 2016-5-27
  * @see {@link OpenXmlIo.XlsxFileType}
@@ -11,14 +11,13 @@ export class XlsxFileType implements OpenXmlIo.XlsxFileType {
     public pathIsTemplate: boolean;
     public pathTemplateToken: string | null;
 
-
     /**
-     * @param schemaUrl: the URL of this file's XML DTD schema
-     * @param contentType: the content/mime type name of this file
-     * @param schemaTarget: the 'target' attribute for this file type used in XLSX files
-     * @param xlsxFilePath: the relative path inside an unzipped XLSX file where this file resides (the path can be a template string that needs a specific sheet number or resource identifier to complete)
-     * @param pathIsTemplate: whether the 'xlsxFilePath' is a template
-     * @param pathTemplateToken: the template token/string to replace in 'xslxFilePath' with a sheet number or resource identifier to make it a valid path
+     * @param schemaUrl the URL of this file's XML DTD schema
+     * @param contentType the content/mime type name of this file
+     * @param schemaTarget the 'target' attribute for this file type used in XLSX files
+     * @param xlsxFilePath the relative path inside an unzipped XLSX file where this file resides (the path can be a template string that needs a specific sheet number or resource identifier to complete)
+     * @param pathIsTemplate whether the 'xlsxFilePath' is a template
+     * @param pathTemplateToken the template token/string to replace in 'xslxFilePath' with a sheet number or resource identifier to make it a valid path
      */
     constructor(schemaUrl: string, contentType: string, schemaTarget: string, xlsxFilePath: string, pathIsTemplate: false, pathTemplateToken: string | null);
     constructor(schemaUrl: string, contentType: string, schemaTarget: string, xlsxFilePath: string, pathIsTemplate: boolean, pathTemplateToken: string);
@@ -31,10 +30,8 @@ export class XlsxFileType implements OpenXmlIo.XlsxFileType {
         this.pathTemplateToken = pathTemplateToken;
     }
 
-
     public static getXmlFilePath(sheetNum: string | number | null, info: OpenXmlIo.XlsxFileType) {
         // TODO the path template token may not be a sheet number, could be a resource identifier (i.e. an image or item prop number)
         return (info.pathIsTemplate ? info.xlsxFilePath.split(<string><any>info.pathTemplateToken).join(<string><any>sheetNum) : info.xlsxFilePath);
     }
-
 }